Added a more elaborated example of clustering

+++ sansa-rdf-spark/src/test/java/net/sansa_stack/rdf/spark/mapper/ClusterEntry.java
@@ -15,6 +15,6 @@ public interface ClusterEntry
     ClusterEntry setItem(Resource res);
 
     @IriNs("eg")
-    BigDecimal getValue();
-    ClusterEntry setValue(BigDecimal value);
+    Integer getValue();
+    ClusterEntry setValue(Integer value);
 }
